Easing functions + doubleToInt + refactoring

+++ AnimationConfig/AnimationUtils.cs
@@ -2,22 +2,22 @@ namespace Time.AnimationConfig;
 using Time.Components;
 public class AnimationUtils
 {
-    public static double ArmStateToDegree(ArmState state)
+    public static int ArmStateToDegree(ArmState state)
     {
         switch (state)
         {
             case ArmState.Zero:
-                return 0.0;
+                return 0;
             case ArmState.Three:
-                return 90.0;
+                return 90;
             case ArmState.Six:
-                return 180.0;
+                return 180;
             case ArmState.Nine:
-                return 270.0;
+                return 270;
             case ArmState.None:
-                return 225.0;
+                return 225;
             case ArmState.HPOne:
-                return 45.0;
+                return 45;
             default: return 0;
         }
     }
